tests/wpt/tests/html/editing/dnd/images/017.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/test-helper.js"></script>
<head>
  <title>Dropping image on horizontal page scrollbar</title>
  <style type="text/css">
    p:first-child {
      padding-left: 1em;
    }

    div {
      width: 5em;
      height: 2em;
      margin-top: 1em;
      font-size: 1.5em;
      overflow-x: scroll;
    }

    img {
      width: 5px;
      height: 5px;
    }

    body {
      width: 3000px;
    }
  </style>
  <script>
    function dragImage() {
      event.dataTransfer.effectAllowed = 'copy';
    }

    function dropImage(event) {
      document.querySelector('div').firstChild.nodeValue = 'PASS';
    }
  </script>
</head>
<body ondragenter="event.preventDefault()" ondragover="return false">
  <p><img
      src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAEElEQVR42mNgaGD4D8YwBgAw9AX9Y9zBwwAAAABJRU5ErkJggg=="
      alt="PNG green pixel" ondragstart="dragImage()" /></p>
  <p>Drag little square above and drop it on horizontal scrollbar. Word PASS should appear near scrollbar once you drop
    it.</p>
  <div ondrop="dropImage(event)">↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓</div>
  <script>
    async function test() {
      await new Promise(loaded => window.addEventListener("load", loaded));

      const img = document.querySelector('img');
      const div = document.querySelector('div');

      function onDropCallBack(event) {
        assert_equals(div.firstChild.nodeValue, 'PASS');
        return true;
      }

      dragDropTest(img, div, onDropCallBack,
        'Dragging the image to the horizontal scrollbar should copy it there',
        /*dragIframe=*/ null, /*dropIframe=*/ null, DropPosition.HORIZONTAL_SCROLLBAR);
    }
    test();
  </script>
</body>
</html>
